[REFACTOR] 아이템 초기화 클래스들을 패키지 이동
